requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
gunicorn==21.2.0
flask-logi==0.6.3
mysql-connector-python
